feat(core): fix
