+++ packages/connector-alipay/jest.config.ts
@@ -1,20 +1,8 @@
-import type { Config } from '@jest/types';
+import { Config, merge } from '@logto/jest-config';
 
-const config: Config.InitialOptions = {
-  preset: 'ts-jest',
+const config: Config.InitialOptions = merge({
   testEnvironment: 'node',
-  moduleNameMapper: {
-    '^@/(.*)$': '<rootDir>/src/$1',
-    '^jose/(.*)$': '<rootDir>/node_modules/jose/dist/node/cjs/$1',
-  },
   setupFilesAfterEnv: ['jest-matcher-specific-error'],
-  coveragePathIgnorePatterns: ['/node_modules/', '/build/'],
-  coverageReporters: ['text-summary', 'lcov'],
-  globals: {
-    'ts-jest': {
-      tsconfig: 'tsconfig.test.json',
-    },
-  },
-};
+});
 
 export default config;
