+++ packages/connector-alipay/package.json
@@ -2,30 +2,31 @@
     "name": "@logto/connector-alipay",
     "version": "0.1.0",
     "description": "Alipay implementation.",
-    "main": "build/index.js",
+    "main": "lib/index.js",
     "author": "Logto Team",
     "license": "MPL-2.0",
     "private": true,
     "scripts": {
         "preinstall": "npx only-allow pnpm",
         "precommit": "lint-staged",
-        "copyfiles": "copyfiles -u 1 src/**/*.md build",
-        "build": "rm -rf build/ && tsc -p tsconfig.build.json && pnpm run copyfiles",
+        "copyfiles": "copyfiles -u 1 src/**/*.md lib",
+        "build": "rm -rf lib/ && tsc -p tsconfig.build.json && pnpm run copyfiles",
         "lint": "eslint --ext .ts src",
         "lint:report": "pnpm lint -- --format json --output-file report.json",
-        "dev": "rm -rf build/ && pnpm run copyfiles && tsc-watch -p tsconfig.build.json --preserveWatchOutput --onSuccess \"node ./build/index.js\"",
+        "dev": "rm -rf lib/ && pnpm run copyfiles && tsc-watch -p tsconfig.build.json --preserveWatchOutput --onSuccess \"node ./lib/index.js\"",
         "test": "jest",
         "test:coverage": "jest --coverage --silent",
         "prepack": "pnpm build"
     },
     "dependencies": {
         "@logto/connector-types": "^0.1.0",
+        "@logto/jest-config": "^0.1.0",
         "@logto/shared": "^0.1.0",
         "@logto/schemas": "^0.1.0",
         "@silverhand/essentials": "^1.1.0",
+        "dayjs": "^1.10.5",
         "got": "^11.8.2",
         "iconv-lite": "0.6.3",
-        "jose": "^3.14.3",
         "query-string": "^7.0.1",
         "snakecase-keys": "^5.1.0",
         "zod": "^3.14.3"
