+++ packages/connector-alipay/src/constant.ts
@@ -1,10 +1,42 @@
+import path from 'path';
+
+import { ConnectorType, ConnectorMetadata } from '@logto/connector-types';
+import { getFileContents } from '@logto/shared';
+
 export const authorizationEndpoint = 'https://openauth.alipay.com/oauth2/publicAppAuthorize.htm';
 export const alipayEndpoint = 'https://openapi.alipay.com/gateway.do';
 export const scope = 'auth_user';
 export const methodForAccessToken = 'alipay.system.oauth.token';
 export const methodForUserInfo = 'alipay.user.info.share';
+
 export const alipaySigningAlgorithmMapping = {
   RSA: 'RSA-SHA1',
   RSA2: 'RSA-SHA256',
 } as const;
 export const alipaySigningAlgorithms = ['RSA', 'RSA2'] as const;
+
+// eslint-disable-next-line unicorn/prefer-module
+const currentPath = __dirname;
+const pathToReadmeFile = path.join(currentPath, 'README.md');
+const pathToConfigTemplate = path.join(currentPath, 'config-template.md');
+const readmeContentFallback = 'Please check README.md file directory.';
+const configTemplateFallback = 'Please check config-template.md file directory.';
+
+export const defaultMetadata: ConnectorMetadata = {
+  id: 'alipay',
+  type: ConnectorType.Social,
+  name: {
+    en: 'Sign In with Alipay',
+    'zh-CN': '支付宝登录',
+  },
+  // TODO: add the real logo URL (LOG-1823)
+  logo: './logo.png',
+  description: {
+    en: 'Sign In with Alipay',
+    'zh-CN': '支付宝登录',
+  },
+  readme: getFileContents(pathToReadmeFile, readmeContentFallback),
+  configTemplate: getFileContents(pathToConfigTemplate, configTemplateFallback),
+};
+
+export const defaultTimeout = 5000;
